fix(calendar styles): create a new layout to select types of calendar
